LGPL/acsBUILD/src/alma/acs/releasedoc/Cvs2clXmlParser.java
package alma.acs.releasedoc;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.FileReader;
import java.io.IOException;

import javax.xml.parsers.DocumentBuilder;
import javax.xml.parsers.DocumentBuilderFactory;
import javax.xml.parsers.ParserConfigurationException;

import org.w3c.dom.Document;
import org.xml.sax.InputSource;
import org.xml.sax.SAXException;

public class Cvs2clXmlParser
{

	
	public Document parseXml(File xmlFile) throws ParserConfigurationException, FileNotFoundException, SAXException, IOException {
		DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
		factory.setNamespaceAware(false);
		DocumentBuilder builder = factory.newDocumentBuilder();

		// parse the XML file into a DOM
		Document parentDoc = builder.parse(new InputSource(new FileReader(xmlFile)));

		return parentDoc;
	}
}
